# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SnapPath AI Brain API",
    "version": "1.1.0",
    "description": "SnapPath API service documentation untuk client terdaftar. Client eksternal memakai x-api-key dan x-api-secret, sedangkan dashboard internal SnapPath memakai JWT session user login."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Public API key client, contoh: sp_xxx"
      },
      "ApiSecretAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-secret",
        "description": "API secret client, contoh: sps_xxx"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Alternative auth: Authorization Basic base64(apiKey:apiSecret)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "QueuedJobResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "runId": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "statusUrl": {
            "type": "string",
            "example": "/api/v1/jobs/{jobId}/status"
          }
        }
      },
      "ClaimValidationInput": {
        "type": "object",
        "required": [
          "providerId",
          "patient",
          "encounter",
          "diagnoses",
          "procedures",
          "medications",
          "totalClaimAmount"
        ],
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Optional. Jika memakai API key client, clientId otomatis diambil dari credential."
          },
          "providerId": {
            "type": "string",
            "format": "uuid",
            "description": "Provider klaim/tarif yang digunakan untuk threshold dan master tariff."
          },
          "claimId": {
            "type": "string",
            "example": "CLM-2026-0001"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "encounter": {
            "$ref": "#/components/schemas/Encounter"
          },
          "diagnoses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diagnosis"
            }
          },
          "procedures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Procedure"
            }
          },
          "medications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Medication"
            }
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClinicalDocument"
            }
          },
          "totalClaimAmount": {
            "type": "number",
            "example": 3250000
          },
          "currency": {
            "type": "string",
            "default": "IDR"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "P-001"
          },
          "name": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "1985-04-21"
          },
          "gender": {
            "type": "string",
            "enum": [
              "M",
              "F"
            ]
          }
        }
      },
      "Encounter": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "RAWAT_INAP",
              "RAWAT_JALAN",
              "IGD"
            ]
          },
          "admissionDate": {
            "type": "string",
            "format": "date-time"
          },
          "dischargeDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "facility": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "example": "RS_TIPE_B"
              }
            }
          }
        }
      },
      "Diagnosis": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "J18.9"
          },
          "description": {
            "type": "string",
            "example": "Pneumonia, unspecified organism"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "COMPLICATION"
            ]
          },
          "sequence": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "Procedure": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "RAD-CXR"
          },
          "description": {
            "type": "string",
            "example": "Chest X-Ray"
          },
          "quantity": {
            "type": "integer",
            "example": 1
          },
          "unitPrice": {
            "type": "number",
            "example": 250000
          },
          "totalPrice": {
            "type": "number",
            "example": 250000
          },
          "performedDate": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string",
            "example": "RADIOLOGI"
          },
          "regionCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Medication": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Azithromycin 500mg"
          },
          "genericName": {
            "type": "string",
            "nullable": true
          },
          "dosage": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "example": 5
          },
          "unitPrice": {
            "type": "number",
            "example": 15000
          },
          "totalPrice": {
            "type": "number",
            "example": 75000
          },
          "frequency": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ClinicalDocument": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "RESUME_MEDIS"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ClinicalPathwayInput": {
        "type": "object",
        "required": [
          "diagnosisCode",
          "encounterType"
        ],
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "providerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "diagnosisCode": {
            "type": "string",
            "example": "A09"
          },
          "diagnosisName": {
            "type": "string",
            "example": "Infectious gastroenteritis and colitis"
          },
          "encounterType": {
            "type": "string",
            "enum": [
              "RAWAT_INAP",
              "RAWAT_JALAN",
              "IGD"
            ]
          },
          "providerType": {
            "type": "string",
            "nullable": true,
            "example": "BPJS"
          },
          "patientProfile": {
            "type": "object",
            "properties": {
              "age": {
                "type": "integer"
              },
              "gender": {
                "type": "string",
                "enum": [
                  "M",
                  "F"
                ]
              },
              "comorbidities": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "includeCosting": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ClinicalPathwayOutput": {
        "type": "object",
        "description": "Core result clinical pathway. Jika dipanggil via /pathways/generate, result ini tersedia di GET /api/v1/jobs/{jobId}/result pada field result.",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "diagnosisCode": {
            "type": "string",
            "example": "A09"
          },
          "diagnosisName": {
            "type": "string"
          },
          "pathwayVersion": {
            "type": "string",
            "example": "1.0-ai"
          },
          "estimatedLos": {
            "type": "integer",
            "description": "Perkiraan Length of Stay standar dalam hari."
          },
          "phases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClinicalPathwayPhase"
            }
          },
          "totalEstimatedCost": {
            "type": "number",
            "nullable": true
          },
          "generatedBy": {
            "type": "string",
            "enum": [
              "AI",
              "TEMPLATE",
              "HYBRID"
            ]
          },
          "confidence": {
            "type": "number",
            "description": "Confidence internal terhadap pathway yang dihasilkan."
          }
        }
      },
      "ClinicalPathwayPhase": {
        "type": "object",
        "description": "Satu fase perjalanan perawatan. Phase biasanya Admission/IGD, Day 1-2, Day 3-5, Discharge.",
        "properties": {
          "phaseId": {
            "type": "string",
            "example": "admission"
          },
          "phaseName": {
            "type": "string",
            "example": "IGD/Admisi"
          },
          "dayRange": {
            "type": "string",
            "example": "Day 0"
          },
          "objectives": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Target klinis fase ini."
          },
          "assessments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PathwayAssessment"
            }
          },
          "treatments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PathwayTreatment"
            }
          },
          "medications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PathwayMedication"
            }
          },
          "nursing": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PathwayNursing"
            }
          },
          "nutrition": {
            "$ref": "#/components/schemas/PathwayNutrition"
          },
          "education": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dischargeGate": {
            "$ref": "#/components/schemas/DischargeGate"
          }
        }
      },
      "PathwayAssessment": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "frequency": {
            "type": "string"
          },
          "mandatory": {
            "type": "boolean"
          }
        }
      },
      "PathwayTreatment": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "route": {
            "type": "string",
            "nullable": true
          },
          "mandatory": {
            "type": "boolean"
          }
        }
      },
...
      "PathwayNursing": {
        "type": "object",
        "properties": {
          "activity": {
            "type": "string"
          },
          "frequency": {
            "type": "string"
          }
        }
      },
...
      "ClaimValidationOutput": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "VALID",
              "INVALID",
              "WARNING",
              "REVIEW_NEEDED"
            ]
          },
          "overallScore": {
            "type": "integer",
            "description": "Skor akhir 0-100 setelah scoring deduction."
          },
          "summary": {
            "type": "string"
          },
          "scoringExplanation": {
            "$ref": "#/components/schemas/ScoringExplanation"
          },
          "documentValidation": {
            "type": "object"
          },
          "diagnosisValidation": {
            "type": "object"
          },
          "tariffValidation": {
            "type": "object"
          },
          "drugPriceValidation": {
            "type": "object"
          },
          "clinicalPathway": {
            "$ref": "#/components/schemas/ClaimClinicalPathwaySection"
          },
          "processingTime": {
            "type": "object"
          },
          "auditTrail": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ScoringExplanation": {
        "type": "object",
        "description": "Penjelasan scoring yang digunakan di SnapPath UI. Skor dimulai dari 100 lalu dikurangi berdasarkan temuan validasi.",
        "properties": {
          "baseScore": {
            "type": "integer",
            "example": 100
          },
          "deductions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "parameter": {
                  "type": "string",
                  "example": "Diagnosis & tindakan klinis"
                },
                "maxDeduction": {
                  "type": "integer",
                  "example": 30
                },
                "trigger": {
                  "type": "string",
                  "example": "diagnosisValidation.isValid === false"
                },
                "meaning": {
                  "type": "string"
                }
              }
            },
            "example": [
              {
                "parameter": "Diagnosis & tindakan klinis",
                "maxDeduction": 30,
                "trigger": "Diagnosis tidak sesuai tindakan atau prosedur wajib hilang",
                "meaning": "Mendeteksi mismatch diagnosis-treatment, prosedur tidak relevan, atau prosedur wajib belum ada."
              },
              {
                "parameter": "Tarif tindakan",
                "maxDeduction": 20,
                "trigger": "tariffValidation.status WARNING/INVALID",
                "meaning": "Ada item tarif di luar master tariff atau melewati threshold."
              },
              {
                "parameter": "Harga obat",
                "maxDeduction": 20,
                "trigger": "drugPriceValidation.status WARNING/INVALID",
                "meaning": "Harga obat melewati referensi pasar + threshold atau obat tidak ditemukan."
              },
              {
                "parameter": "Kelengkapan dokumen",
                "maxDeduction": 10,
                "trigger": "documentValidation.isValid === false",
                "meaning": "Dokumen mandatory seperti resume medis, hasil lab/radiologi, rujukan, atau laporan operasi belum lengkap."
              }
            ]
          }
        }
      },
      "ClaimClinicalPathwaySection": {
        "type": "object",
        "description": "Section clinicalPathway di full claim validation. Berisi recommendedPathway hasil AI/template untuk diagnosis primer.",
        "properties": {
          "diagnosisCode": {
            "type": "string"
          },
          "adherenceScore": {
            "type": "integer",
            "example": 100,
            "description": "Saat ini default 100 karena sistem menghasilkan pathway rekomendasi; deviasi detail akan diisi saat adherence engine aktif."
          },
          "recommendedPathway": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClinicalPathwayPhase"
            }
          },
          "deviations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ProviderInput": {
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "code": {
            "type": "string",
            "example": "BPJS"
          },
          "name": {
            "type": "string",
            "example": "BPJS Kesehatan"
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": [],
      "ApiSecretAuth": []
    },
    {
      "BasicAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/claims/map-json": {
      "post": {
        "tags": [
          "Claim Workflows"
        ],
        "summary": "Map arbitrary JSON to SnapPath claim schema",
        "description": "Mengubah payload dari SIMRS/FHIR/HL7/custom export menjadi struktur ClaimValidationInput. Endpoint ini menggunakan AI dan mengembalikan usage token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mapped claim payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "mapped": {
                      "$ref": "#/components/schemas/ClaimValidationInput"
                    },
                    "_mappingNotes": {
                      "type": "string"
                    },
                    "usage": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/validate": {
      "post": {
        "tags": [
          "Claim Workflows"
        ],
        "summary": "Full Claim Validation",
        "description": "Menjalankan workflow asynchronous: document validation, diagnosis-treatment validation, tariff validation, drug price check, dan clinical pathway generation untuk diagnosis primer.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimValidationInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedJobResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{jobId}/status": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Check job status",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job metadata and status."
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/result": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get completed job result",
        "description": "Mengambil result JSON final. Untuk PATHWAY_GEN, field result mengikuti ClinicalPathwayOutput. Untuk CLAIM_VALIDATION, field result mengikuti ClaimValidationOutput.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed job result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "jobType": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "result": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/ClinicalPathwayOutput"
                        },
                        {
                          "$ref": "#/components/schemas/ClaimValidationOutput"
                        }
                      ]
                    },
                    "completedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Job not finished yet"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/v1/pathways/generate": {
      "post": {
        "tags": [
          "Clinical Pathway"
        ],
        "summary": "Generate Clinical Pathway",
        "description": "Generate pathway phase-by-phase untuk diagnosis tertentu. Result final diambil melalui GET /api/v1/jobs/{jobId}/result.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicalPathwayInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedJobResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/providers": {
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Create claim/tariff provider",
        "description": "Membuat provider klaim/tarif di bawah client. Endpoint ini membutuhkan JWT session internal, bukan API key client eksternal.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}